latest commit

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -4,12 +4,22 @@ namespace App\Http\Controllers\Admin;
 
 use App\Data\Admin\Category\CategoryData;
 use App\Data\Admin\Category\CreateCategoryData;
-use App\Data\Admin\Category\ParentCategoryIdsData;
+use App\Data\Admin\Category\PaginatedCategoryData;
 use App\Data\Admin\Category\PathParameters\CategoryIdPathParameterData;
+use App\Data\Admin\Category\QueryParameters\CategoryIdsQueryParameterData;
+use App\Data\Admin\Category\QueryParameters\Index\CategoryIndexQueryParameterData;
 use App\Data\Admin\Category\UpdateCategoryData;
+use App\Data\Shared\ListData;
+use App\Data\Shared\Swagger\Parameter\QueryParameter\ListQueryParameter;
+use App\Data\Shared\Swagger\Parameter\QueryParameter\QueryParameter;
+use App\Data\Shared\Swagger\Request\FormDataRequestBody;
+use App\Data\Shared\Swagger\Response\SuccessItemResponse;
+use App\Data\Shared\Swagger\Response\SuccessListResponse;
+use App\Data\Shared\Swagger\Response\SuccessNoContentResponse;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
 use App\Services\FileService;
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Support\Facades\Log;
 use OpenApi\Attributes as OAT;
 
@@ -50,66 +60,75 @@ class CategoryController extends Controller
     /**
      * Get All Categories
      */
-    #[OAT\Get(
-        path: '/admin/categories',
-        tags: ['categories'],
-        //        parameters: [
-        //            new OAT\QueryParameter(
-        //                required: false,
-        //                ref: "#/components/parameters/categorySearch",
-        //            ),
-        //            new OAT\QueryParameter(
-        //                required: false,
-        //                ref: "#/components/parameters/categoryOrderBy",
-        //            ),
-        //            new OAT\QueryParameter(
-        //                required: false,
-        //                ref: "#/components/parameters/categoryDirection"
-        //            ),
-        //            new OAT\QueryParameter(
-        //                required: false,
-        //                ref: "#/components/parameters/categoryPerPage"
-        //            ),
-        //            new OAT\QueryParameter(
-        //                required: false,
-        //                ref: "#/components/parameters/categoryPage"
-        //            ),
-        //        ],
-        responses: [
-            new OAT\Response(
-                response: 200,
-                description: 'The Category was successfully created',
-                //                content: new OAT\JsonContent(ref: '#/components/schemas/paginatedCategory'),
-                content: new OAT\JsonContent(
-                    type: 'array',
-                    items: new OAT\Items(
-                        type: CategoryData::class
-                    ),
-                ),
-            ),
-        ],
-    )]
-    public function index()
-    {
+    #[OAT\Get(path: '/admin/categories', tags: ['categories'])]
+    #[QueryParameter('page', 'integer')]
+    #[QueryParameter('perPage', 'integer')]
+    #[QueryParameter('search')]
+    #[ListQueryParameter]
+    #[QueryParameter('sort')]
+    #[SuccessItemResponse(PaginatedCategoryData::class)]
+    public function index(
+        CategoryIndexQueryParameterData $query_parameters
+    ) {
 
-        //        Log::info('accessing CategoryController index method');
-        //
-        //                Log::info('categories parent_ids {categories} ', ['categories' => Category::pluck('parent_id')]);
+        Log::info('accessing Admin CategoryController index method');
 
-        $categoriesData = CategoryData::collect(
-            Category::all()->select([
-                'id',
-                'parent_id',
-                'name',
-                'image',
-                'created_at',
+        $search_filter = $query_parameters->search;
+
+        $parent_ids_filter = $query_parameters->ids;
+
+        Log::info($query_parameters);
+
+        $per_page = $query_parameters->perPage ?? 10;
+
+        $sort = $query_parameters->sort;
+
+        Log::info($per_page);
+
+        $categories = Category::query()
+            ->leftJoin(
+                'categories as parent',
+                'categories.parent_id',
+                '=',
+                'parent.id',
+            )// can now reference related table(same table with sub-query) using parent.$column
+            // in subsequent elequent methods
+            ->when($search_filter, function (Builder $query) use ($search_filter) {
+                return $query->whereAnyLike(
+                    ['categories.id', 'categories.name', 'categories.created_at'],
+                    $search_filter
+                );
+            })
+            ->when($parent_ids_filter, function (Builder $query) use ($parent_ids_filter) {
+
+                return $query->whereIn(
+                    'categories.parent_id',
+                    $parent_ids_filter
+                );
+            })
+            ->when($sort, function (Builder $query) use ($sort) {
+
+                [$sort_field, $sort_value] = explode(' ', $sort);
+
+                return $query->orderByDynamic($sort_field, $sort_value);
+            })
+            ->select([
+                'categories.id',
+                'categories.parent_id',
+                'categories.name',
+                'categories.image',
+                'categories.created_at',
+                'parent.name as parent_name',
             ])
-        );
+            ->paginate(perPage: $per_page);
 
-        Log::info(
-            'Fetched categories {categories}',
-            ['categories' => $categoriesData]
-        );
+        //        Log::info($categories);
+
+        return $categories;
+
+        $categoriesData = CategoryData::collect($categories);
+
+        Log::info($categoriesData);
 
         return $categoriesData;
     }
@@ -117,42 +136,12 @@ class CategoryController extends Controller
     /**
      * Create a new Category.
      */
-    #[OAT\Post(
-        path: '/admin/categories',
-        requestBody: new OAT\RequestBody(
-            required: true,
-            content: new OAT\MediaType(
-                mediaType: 'multipart/form-data',
-                schema: new OAT\Schema(
-                    type: CreateCategoryData::class,
-                ),
-            ),
-        ),
-        tags: ['categories'],
-        parameters: [
-
-        ],
-        responses: [
-            // new OAT\Response(
-            //     response: 200,
-            //     description: 'Category Created Successfully',
-            //     content: new OAT\JsonContent(type: CategoryData::class),
-            // ),
-            new OAT\Response(
-                response: 200,
-                description: 'Category was created successfully',
-                content: new OAT\JsonContent(
-                    type: 'array',
-                    items: new OAT\Items(
-                        type: CategoryData::class,
-                    ),
-                ),
-            ),
-        ],
-    )]
+    #[OAT\Post(path: '/admin/categories', tags: ['categories'])]
+    #[FormDataRequestBody(CreateCategoryData::class)]
+    #[SuccessNoContentResponse('Category created successfully')]
     public function store(
         CreateCategoryData $createCategoryData,
-    ): CategoryData {
+    ) {
 
         Log::info('Accessing CategoryController store method');
 
@@ -168,37 +157,36 @@ class CategoryController extends Controller
 
         Log::info('Category was created {category}', ['category' => $category]);
 
-        return CategoryData::from($category);
     }
 
-    #[OAT\Get(
-        path: '/admin/categories/{id}',
-        tags: ['categories'],
-        parameters: [
+    /** Get sub Categories by the id of the category */
+    #[OAT\Get(path: '/admin/categories/list', tags: ['categories'])]
+    #[SuccessNoContentResponse('Category list fetched successfully')]
+    public function list()
+    {
+        Log::info(
+            'accessing Admin CategoryController, list method',
+        );
 
-        ],
-        responses: [
-            new OAT\Response(
-                response: 200,
-                description: 'category fetched successfully',
-                content: new OAT\JsonContent(type: CategoryData::class),
-            ),
-        ],
-    )]
+        $categories =
+            Category::query()
+                ->select(['id', 'name'])
+                ->get();
+
+        return ListData::collect($categories);
+    }
+
+    #[OAT\Get(path: '/admin/categories/{id}', tags: ['categories'])]
+    #[SuccessNoContentResponse('Fetched category successfully')]
     public function show(CategoryIdPathParameterData $request)
     {
 
         Log::info(
-            'accessing Category Controller show method with path id {id}',
+            'accessing Admin CategoryController show method with path id {id}',
             ['id' => $request->id]
         );
 
         $category = Category::find($request->id);
-
-        Log::info(
-            'Fetched category {category}',
-            ['category' => $category]
-        );
 
         return CategoryData::from($category);
     }
@@ -206,69 +194,40 @@ class CategoryController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    #[OAT\Patch(
-        path: '/admin/categories/{id}',
-        requestBody: new OAT\RequestBody(
-            required: true,
-            content: new OAT\MediaType(
-                mediaType: 'multipart/form-data',
-                schema: new OAT\Schema(
-                    type: UpdateCategoryData::class,
-                ),
-            ),
-        ),
-        tags: ['categories'],
-        parameters: [
-
-        ],
-        responses: [
-            new OAT\Response(
-                response: 200,
-                description: 'Category was updated successfully',
-                content: new OAT\JsonContent(
-                    type: 'array',
-                    items: new OAT\Items(
-                        type: CategoryData::class,
-                    ),
-                ),
-            ),
-        ],
-    )]
+    #[OAT\Patch(path: '/admin/categories/{id}', tags: ['categories'])]
+    #[FormDataRequestBody(UpdateCategoryData::class)]
+    #[SuccessNoContentResponse('Category was updated successfully')]
     public function update(
         CategoryIdPathParameterData $request,
         UpdateCategoryData $updatedCategoryData,
     ) {
+
+        Log::info('accessing Admin CategoryController update method');
+
+        Log::info($updatedCategoryData);
+
         $category = Category::find($request->id);
 
         $isCategoryUpdated = $category->update([
-            'name' => $request->name,
-            'image' => $request->image,
+            'name' => $updatedCategoryData->name,
+            'parent_id' => $updatedCategoryData->parent_id,
+            'image' => $updatedCategoryData->image,
         ]);
 
-        if ($isCategoryUpdated) {
+        if ($isCategoryUpdated && $category->image) {
             FileService::delete('category', $category->image);
         }
-
-        return CategoryData::from($category);
 
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    #[OAT\Delete(
-        path: '/admin/categories/{id}',
-        tags: ['categories'],
-        responses: [
-            new OAT\Response(
-                response: 204,
-                description: 'The Category was successfully deleted',
-            ),
-        ],
-    )]
-    public function destroy(CategoryIdPathParameterData $request): bool
+    #[OAT\Delete(path: '/admin/categories/{id}', tags: ['categories'])]
+    #[SuccessNoContentResponse('Item Deleted Successfully')]
+    public function destroy(CategoryIdPathParameterData $path_parameters): bool
     {
-        $categoryToDelete = Category::find($request->id);
+        $categoryToDelete = Category::find($path_parameters->id);
 
         $isCategoryDeleted = $categoryToDelete->delete();
 
@@ -276,21 +235,25 @@ class CategoryController extends Controller
 
     }
 
-    /** Get sub Categories by the id of the category */
-    #[OAT\Get(
-        path: '/admin/categories/getSubCategories/{id}',
-        tags: ['categories'],
-        parameters: [
+    /** Get Parent Categories List*/
+    #[OAT\Get(path: '/admin/categories/parentList', tags: ['categories'])]
+    #[SuccessListResponse(ListData::class)]
+    public function getParentCategoriesList()
+    {
+        Log::info(
+            'accessing Category Controller, getParentCategories method',
+        );
 
-        ],
-        responses: [
-            new OAT\Response(
-                response: 200,
-                description: 'category fetched successfully',
-                content: new OAT\JsonContent(type: CategoryData::class),
-            ),
-        ],
-    )]
+        $parent_categories = Category::query()
+            ->isParent()
+            ->get();
+
+        return ListData::collect($parent_categories);
+    }
+
+    /** Get Child Categories List */
+    #[OAT\Get(path: '/admin/categories/getSubCategories/{id}', tags: ['categories'])]
+    #[SuccessItemResponse(CategoryData::class)]
     public function getSubCategories(CategoryIdPathParameterData $request)
     {
         Log::info(
@@ -308,34 +271,20 @@ class CategoryController extends Controller
     }
 
     /** Get child Categories by parent Ids List */
-    #[OAT\Post(
-        path: '/admin/categories/GetSubCategoriesByParents',
-        requestBody: new OAT\RequestBody(
-            required: true,
-            content: new OAT\JsonContent(type: ParentCategoryIdsData::class),
-        ),
-        tags: ['categories'],
-        responses: [
-            new OAT\Response(
-                response: 200,
-                description: 'Fetched Categories successfully',
-                content: new OAT\JsonContent(
-                    type: 'array',
-                    items: new OAT\Items(
-                        type: CategoryData::class,
-                    ),
-                ),
-            ),
-        ],
-    )]
-    public function GetSubCategoriesByParents(ParentCategoryIdsData $request)
+    #[OAT\Get(path: '/admin/categories/GetSubCategoriesByParents', tags: ['categories'])]
+    #[QueryParameter('name')]
+    #[ListQueryParameter]
+    #[SuccessListResponse(CategoryData::class)]
+    public function GetSubCategoriesByParents(CategoryIdsQueryParameterData $query_parameters)
     {
 
-        $parentIds = $request->ids;
+        Log::info('accessing Admin CategoryController GetSubCategoriesByParents method');
+
+        $parentIds = $query_parameters->ids;
 
         Log::info(
             'accessing CategoryController, GetSubCategoriesByParents Method with ids  {ids}',
-            ['ids' => $request->ids]
+            ['ids' => $query_parameters->ids]
         );
 
         $childCategories =
